+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Version 0.0.4
 -------------
 
 - Adding more filters for group, account, and QoS
+- Adding sorting capability for running jobs
+- Adding flag for showing nodes in a job
+- Adding flag for displaying job names instead of IDs
 
 Version 0.0.3
 -------------